requirements.txt
requests
pywebcopy
numpy
pandas
mitmproxy
numba
lxml
finplot
plotly
seaborn
matplotlib
scikit-learn
investpy
yfinance
tensorflow-gpu
xgboost
imblearn
setuptools
multitasking
telebot
Telethon
beautifulsoup4
afterhours
keras
textblob
scipy
torch
nltk
transformers
tqdm
scikeras
flair
jsonpickle
stocktrends